refactor(frontend): extract card-render.js from the card
