+++ custom_components/cover_time_based/frontend/cover-time-based-card.js
@@ -8,28 +8,19 @@
  * All user-visible strings are translatable. Translations are embedded below.
  */
 
-import {
-  LitElement,
-  html,
-} from "https://unpkg.com/lit-element@2.4.0/lit-element.js?module";
+import { LitElement } from "https://unpkg.com/lit-element@2.4.0/lit-element.js?module";
 import {
   filterEntitiesByValidEntries,
-  switchPickerDomains,
   switchLabelKey,
-  showsPulseTime,
   clearedEntitiesForMode,
   clearedTiltConfig,
   coverHasNativeTilt,
   coverConfirmedWithoutTilt,
 } from "./entity-filter.js";
-import { renderTextfield } from "./textfield-render.js";
-import {
-  DOMAIN,
-  TIMING_ATTRIBUTES,
-  ATTRIBUTE_TO_CONFIG,
-} from "./constants.js";
+import { DOMAIN, ATTRIBUTE_TO_CONFIG } from "./constants.js";
 import { translate } from "./translations.js";
 import { cardStyles } from "./card-styles.js";
+import { renderCard } from "./card-render.js";
 
 class CoverTimeBasedCard extends LitElement {
   static get properties() {
@@ -541,170 +532,10 @@ class CoverTimeBasedCard extends LitElement {
   // --- Rendering ---
 
   render() {
-    if (!this.hass) return html``;
-    return html`
-      <ha-card>
-        <div class="card-header">${this._t("header")}</div>
-        <div class="card-content">
-          ${this._renderEntityPicker()}
-          ${this._selectedEntity && this._config
-            ? this._renderConfigSections()
-            : ""}
-          ${this._loadError
-            ? html`<div class="yaml-warning">${this._loadError}</div>`
-            : ""}
-          ${this._loading
-            ? html`<div class="loading">
-                <ha-icon icon="mdi:loading" class="spin"></ha-icon> ${this._t("loading")}
-              </div>`
-            : ""}
-        </div>
-        ${this._openHelp
-          ? html`<div
-              class="popover-backdrop"
-              @click=${this._closeHelp}
-            ></div>`
-          : ""}
-      </ha-card>
-    `;
+    return renderCard(this);
   }
 
-  _renderEntityPicker() {
-    return html`
-      <div class="section">
-        <ha-entity-picker
-          .hass=${this.hass}
-          .value=${this._selectedEntity}
-          .includeEntities=${this._configEntryEntities || []}
-          label=""
-          @value-changed=${(e) => {
-            const newEntity = e.detail?.value || "";
-            if (newEntity === this._selectedEntity) return;
-            if (this._isCalibrating()) {
-              if (!confirm(this._t("confirm_cancel_calibration"))) {
-                const current = this._selectedEntity;
-                const picker = e.target;
-                picker.value = current;
-                requestAnimationFrame(() => {
-                  picker.value = current;
-                });
-                this.requestUpdate();
-                return;
-              }
-              if (this._isCalibrating()) {
-                this._onStopCalibration(true);
-              }
-            }
-            this._selectedEntity = newEntity;
-            this._config = null;
-            this._loadError = null;
-            this._knownPosition = "unknown";
-            this._calibratingOverride = undefined;
-            this._activeTab = "device";
-            if (this._selectedEntity) this._loadConfig();
-          }}
-        ></ha-entity-picker>
-        <a class="create-new-link" href="#" @click=${(e) => {
-          e.preventDefault();
-          this._onCreateNew();
-        }}>${this._t("create_new")}</a>
-      </div>
-    `;
-  }
 
-  _renderConfigSections() {
-    const c = this._config;
-    const calibrating = this._isCalibrating();
-    const disabled = this._saving || calibrating;
-
-    return html`
-      <div class="entity-info">
-        <div class="entity-info-row">
-          <div>
-            <strong>
-              ${this._getEntityState()?.attributes?.friendly_name ||
-              this._selectedEntity}
-            </strong>
-            <span class="entity-id">${this._selectedEntity}</span>
-          </div>
-        </div>
-      </div>
-
-      <div class="tabs">
-        <button
-          class="tab ${this._activeTab === "device" ? "active" : ""}"
-          @click=${() => { this._activeTab = "device"; }}
-        >${this._t("tabs.device")}</button>
-        <button
-          class="tab ${this._activeTab === "timing" ? "active" : ""}"
-          ?disabled=${!this._hasRequiredEntities(c)}
-          @click=${() => { this._activeTab = "timing"; }}
-        >${this._t("tabs.calibration")}</button>
-      </div>
-
-      ${this._activeTab === "device"
-        ? html`
-            <fieldset ?disabled=${disabled}>
-              ${this._renderControlMode(c)} ${this._renderInputEntities(c)}
-              ${this._renderTiltSupport(c)}
-              ${this._renderTiltMotorSection(c)}
-            </fieldset>
-          `
-        : html`
-            ${calibrating ? "" : this._renderPositionReset()}
-            ${this._renderCalibration(calibrating)}
-            ${this._renderTimingTable(c)}
-          `}
-
-      ${this._saving
-        ? html`<div class="save-bar"><span class="saving-indicator">${this._t("saving")}</span></div>`
-        : ""}
-      ${this._saveError
-        ? html`<div class="save-bar"><span class="save-error">${this._t("save_failed")}</span></div>`
-        : ""}
-    `;
-  }
-
-  _renderControlMode(c) {
-    const mode = c.control_mode || "switch";
-    const showPulseTime = showsPulseTime(mode);
-
-    return html`
-      <div class="section">
-        <div class="field-label">${this._t("control_mode.label")}</div>
-        <select class="ha-select" @change=${this._onControlModeChange}>
-          <option value="wrapped" ?selected=${mode === "wrapped"}>
-            ${this._t("control_mode.wrapped")}
-          </option>
-          <option value="switch" ?selected=${mode === "switch"}>
-            ${this._t("control_mode.switch")}
-          </option>
-          <option value="pulse" ?selected=${mode === "pulse"}>
-            ${this._t("control_mode.pulse")}
-          </option>
-          <option value="toggle" ?selected=${mode === "toggle"}>
-            ${this._t("control_mode.toggle")}
-          </option>
-        </select>
-        ${showPulseTime
-          ? html`
-              <div class="inline-field">
-                ${renderTextfield({
-                  type: "number",
-                  min: "0.1",
-                  max: "10",
-                  step: "0.1",
-                  suffix: "s",
-                  label: this._t("control_mode.pulse_time"),
-                  value: String(c.pulse_time || 1.0),
-                  onChange: this._onPulseTimeChange,
-                })}
-              </div>
-            `
-          : ""}
-      </div>
-    `;
-  }
 
   _toggleHelp(helperKey) {
     this._openHelp = this._openHelp === helperKey ? null : helperKey;
@@ -712,538 +543,6 @@ class CoverTimeBasedCard extends LitElement {
 
   _closeHelp() {
     this._openHelp = null;
-  }
-
-  _renderToggleWithHelp(labelKey, helperKey, checked, onChange) {
-    const open = this._openHelp === helperKey;
-    return html`
-      <div class="toggle-with-help">
-        <span class="toggle-label">${this._t(labelKey)}</span>
-        <span class="help-anchor">
-          <ha-icon
-            class="help-icon"
-            icon="mdi:help-circle-outline"
-            role="button"
-            tabindex="0"
-            aria-label=${this._t("more_info")}
-            aria-expanded=${open ? "true" : "false"}
-            @click=${(e) => {
-              e.stopPropagation();
-              this._toggleHelp(helperKey);
-            }}
-            @keydown=${(e) => {
-              if (e.key === "Enter" || e.key === " ") {
-                e.preventDefault();
-                this._toggleHelp(helperKey);
-              } else if (e.key === "Escape") {
-                this._closeHelp();
-              }
-            }}
-          ></ha-icon>
-          ${open
-            ? html`<div class="info-popover" role="tooltip">
-                ${this._t(helperKey)}
-              </div>`
-            : ""}
-        </span>
-        <ha-switch
-          class="toggle-switch"
-          .checked=${checked}
-          @change=${onChange}
-        ></ha-switch>
-      </div>
-    `;
-  }
-
-  _renderInputEntities(c) {
-    if (c.control_mode === "wrapped") {
-      return html`
-        <div class="section">
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.cover_entity_id || ""}
-            .includeDomains=${["cover"]}
-            .entityFilter=${this._filterNonTimeBased}
-            label=${this._t("entities.cover_entity")}
-            @value-changed=${this._onCoverEntityChange}
-          ></ha-entity-picker>
-          ${this._renderToggleWithHelp(
-            "entities.ignore_reported_position",
-            "entities.ignore_reported_position_helper",
-            !!c.ignore_reported_position,
-            (e) =>
-              this._updateLocal({ ignore_reported_position: e.target.checked }),
-          )}
-          ${this._renderToggleWithHelp(
-            "entities.force_time_based_position",
-            "entities.force_time_based_position_helper",
-            !!c.force_time_based_position,
-            (e) =>
-              this._updateLocal({ force_time_based_position: e.target.checked }),
-          )}
-          ${this._renderToggleWithHelp(
-            "assumed_state.label",
-            "assumed_state.helper",
-            c.assumed_state !== false,
-            (e) => this._updateLocal({ assumed_state: e.target.checked }),
-          )}
-        </div>
-      `;
-    }
-
-    return html`
-      <div class="section">
-        <div class="field-label">${this._switchLabel("entities.switch_entities", c.control_mode)}</div>
-        <div class="entity-grid">
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.open_switch_entity_id || ""}
-            .includeDomains=${switchPickerDomains(c.control_mode)}
-            label=${this._switchLabel("entities.open_switch", c.control_mode)}
-            @value-changed=${(e) =>
-              this._onSwitchEntityChange("open_switch_entity_id", e)}
-          ></ha-entity-picker>
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.close_switch_entity_id || ""}
-            .includeDomains=${switchPickerDomains(c.control_mode)}
-            label=${this._switchLabel("entities.close_switch", c.control_mode)}
-            @value-changed=${(e) =>
-              this._onSwitchEntityChange("close_switch_entity_id", e)}
-          ></ha-entity-picker>
-          ${c.control_mode === "pulse" ? html`
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.stop_switch_entity_id || ""}
-            .includeDomains=${switchPickerDomains(c.control_mode)}
-            label=${this._switchLabel("entities.stop_switch", c.control_mode)}
-            @value-changed=${(e) =>
-              this._onSwitchEntityChange("stop_switch_entity_id", e)}
-          ></ha-entity-picker>
-          ` : ""}
-        </div>
-        ${this._renderToggleWithHelp(
-          "assumed_state.label",
-          "assumed_state.helper",
-          c.assumed_state !== false,
-          (e) => this._updateLocal({ assumed_state: e.target.checked }),
-        )}
-      </div>
-    `;
-  }
-
-  _renderTiltSupport(c) {
-    const tiltMode = c.tilt_mode || "none";
-
-    // Dual-motor tilt on a wrapped cover delegates the tilt commands to the
-    // underlying entity, so it requires that cover to support tilt natively.
-    // Inline and sequential modes drive the main open/close motor, so they
-    // work on any wrapped cover and stay available regardless.
-    const allowDualMotor =
-      c.control_mode !== "wrapped" ||
-      this._coverSupportsNativeTilt(c.cover_entity_id);
-    // The handlers reset dual_motor when it stops being backable, so in normal
-    // UI flow allowDualMotor already covers it. Keep showing it when it is the
-    // stored mode as a safety net for hand-edited configs or a wrapped cover
-    // that is momentarily unavailable (features read as 0) — otherwise the
-    // select would have a selected value missing from its options.
-    const showDualMotor = allowDualMotor || tiltMode === "dual_motor";
-
-    return html`
-      <div class="section">
-        <div class="field-label">${this._t("tilt.label")}</div>
-        <select class="ha-select" @change=${this._onTiltModeChange}>
-          <option value="none" ?selected=${tiltMode === "none"}>
-            ${this._t("tilt.none")}
-          </option>
-          <option value="sequential_close" ?selected=${tiltMode === "sequential_close"}>
-            ${this._t("tilt.sequential_close")}
-          </option>
-          <option value="sequential_open" ?selected=${tiltMode === "sequential_open"}>
-            ${this._t("tilt.sequential_open")}
-          </option>
-          ${showDualMotor
-            ? html`
-                <option value="dual_motor" ?selected=${tiltMode === "dual_motor"}>
-                  ${this._t("tilt.dual_motor")}
-                </option>
-              `
-            : ""}
-          <option value="inline" ?selected=${tiltMode === "inline"}>
-            ${this._t("tilt.inline")}
-          </option>
-        </select>
-        ${tiltMode === "sequential_close" || tiltMode === "dual_motor"
-          ? html`
-              <div class="inline-field">
-                <ha-formfield .label=${this._t("tilt.close_includes_tilt")}>
-                  <ha-switch
-                    .checked=${c.close_includes_tilt !== false}
-                    @change=${(e) =>
-                      this._updateLocal({ close_includes_tilt: e.target.checked })}
-                  ></ha-switch>
-                </ha-formfield>
-              </div>
-            `
-          : ""}
-      </div>
-    `;
-  }
-
-  _renderTiltMotorSection(c) {
-    if (c.tilt_mode !== "dual_motor") return "";
-
-    return html`
-      <div class="section">
-        <div class="field-label">${this._switchLabel("tilt_motor.label", c.control_mode)}</div>
-        ${c.control_mode !== "wrapped" ? html`
-        <div class="entity-grid">
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.tilt_open_switch || ""}
-            .includeDomains=${switchPickerDomains(c.control_mode)}
-            label=${this._switchLabel("tilt_motor.open_switch", c.control_mode)}
-            @value-changed=${(e) =>
-              this._onSwitchEntityChange("tilt_open_switch", e)}
-          ></ha-entity-picker>
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.tilt_close_switch || ""}
-            .includeDomains=${switchPickerDomains(c.control_mode)}
-            label=${this._switchLabel("tilt_motor.close_switch", c.control_mode)}
-            @value-changed=${(e) =>
-              this._onSwitchEntityChange("tilt_close_switch", e)}
-          ></ha-entity-picker>
-          ${c.control_mode === "pulse" ? html`
-          <ha-entity-picker
-            .hass=${this.hass}
-            .value=${c.tilt_stop_switch || ""}
-            .includeDomains=${switchPickerDomains(c.control_mode)}
-            label=${this._switchLabel("tilt_motor.stop_switch", c.control_mode)}
-            @value-changed=${(e) =>
-              this._onSwitchEntityChange("tilt_stop_switch", e)}
-          ></ha-entity-picker>
-          ` : ""}
-        </div>
-        ` : ""}
-        <div class="dual-motor-config">
-          ${renderTextfield({
-            type: "number",
-            min: "0",
-            max: "100",
-            step: "1",
-            label: this._t("tilt_motor.safe_position"),
-            hint: this._t("tilt_motor.safe_position_helper"),
-            value: String(c.safe_tilt_position ?? 100),
-            onChange: (e) => {
-              const v = parseInt(e.target.value);
-              if (!isNaN(v) && v >= 0 && v <= 100) {
-                this._updateLocal({ safe_tilt_position: v });
-              }
-            },
-          })}
-          ${renderTextfield({
-            type: "number",
-            min: "0",
-            max: "100",
-            step: "1",
-            label: this._t("tilt_motor.max_allowed_position"),
-            hint: this._t("tilt_motor.max_allowed_helper"),
-            value:
-              c.max_tilt_allowed_position != null
-                ? String(c.max_tilt_allowed_position)
-                : "",
-            onChange: (e) => {
-              const v = e.target.value.trim();
-              this._updateLocal({
-                max_tilt_allowed_position: v === "" ? null : parseInt(v),
-              });
-            },
-          })}
-        </div>
-      </div>
-    `;
-  }
-
-  _renderTimingRow([labelKey, key, value, min = 0]) {
-    return html`
-      <tr>
-        <td>${this._t(labelKey)}</td>
-        <td class="value-cell">
-          <input
-            type="number"
-            class="timing-input"
-            .value=${value != null ? String(value) : ""}
-            placeholder=${this._t("timing.not_set")}
-            step="0.1"
-            min="${min}"
-            max="600"
-            @change=${(e) => {
-              const v = e.target.value.trim();
-              this._updateLocal({ [key]: v === "" ? null : parseFloat(v) });
-            }}
-          /><span class="unit">s</span>
-        </td>
-      </tr>
-    `;
-  }
-
-  _renderTimingTable(c) {
-    const hasTiltTimes = c.tilt_mode === "sequential_close" || c.tilt_mode === "sequential_open" || c.tilt_mode === "dual_motor" || c.tilt_mode === "inline";
-
-    const travelRows = [
-      ["timing.travel_time_close", "travel_time_close", c.travel_time_close, 0.1],
-      ["timing.travel_time_open", "travel_time_open", c.travel_time_open, 0.1],
-      ["timing.travel_startup_delay", "travel_startup_delay", c.travel_startup_delay],
-      ["timing.min_movement_time", "min_movement_time", c.min_movement_time],
-    ];
-    // Endpoint run-on only applies to switch mode (its latched relay must be
-    // de-energized at the endpoint). Pulse/toggle/wrapped covers self-stop at
-    // their limit switches, so the setting has no effect there.
-    if ((c.control_mode || "switch") === "switch") {
-      travelRows.push([
-        "timing.endpoint_runon_time",
-        "endpoint_runon_time",
-        c.endpoint_runon_time,
-      ]);
-    }
-
-    const tiltRows = [
-      ["timing.tilt_time_close", "tilt_time_close", c.tilt_time_close, 0.1],
-      ["timing.tilt_time_open", "tilt_time_open", c.tilt_time_open, 0.1],
-      ["timing.tilt_startup_delay", "tilt_startup_delay", c.tilt_startup_delay],
-    ];
-
-    return html`
-      <div class="section">
-        <table class="timing-table">
-          <thead>
-            <tr>
-              <th>${this._t("timing.travel_attribute_header")}</th>
-              <th>${this._t("timing.value_header")}</th>
-            </tr>
-          </thead>
-          <tbody>
-            ${travelRows.map((row) => this._renderTimingRow(row))}
-          </tbody>
-        </table>
-        ${hasTiltTimes ? html`
-        <table class="timing-table" style="margin-top: 8px;">
-          <thead>
-            <tr>
-              <th>${this._t("timing.tilt_attribute_header")}</th>
-              <th>${this._t("timing.value_header")}</th>
-            </tr>
-          </thead>
-          <tbody>
-            ${tiltRows.map((row) => this._renderTimingRow(row))}
-          </tbody>
-        </table>
-        ` : ""}
-      </div>
-    `;
-  }
-
-  _renderPositionReset() {
-    const tiltMode = this._config?.tilt_mode || "none";
-    const hasIndependentTilt = tiltMode === "sequential_close" || tiltMode === "sequential_open" || tiltMode === "dual_motor" || tiltMode === "inline";
-
-    return html`
-      <div class="section">
-        <div class="field-label">${this._t("position.label")}</div>
-        <div class="helper-text">
-          ${this._t("position.helper")}
-        </div>
-        ${!this._hasTiltMotor() ? html`
-          <div class="cover-controls">
-            <ha-button title=${this._t("controls.open")} @click=${() => this._onCoverCommand("open_cover")}>
-              <ha-icon icon="mdi:window-shutter-open" style="--mdc-icon-size: 18px;"></ha-icon>
-            </ha-button>
-            <ha-button title=${this._t("controls.stop")} @click=${() => this._onCoverCommand("stop_cover")}>
-              <ha-icon icon="mdi:stop" style="--mdc-icon-size: 18px;"></ha-icon>
-            </ha-button>
-            <ha-button title=${this._t("controls.close")} @click=${() => this._onCoverCommand("close_cover")}>
-              <ha-icon icon="mdi:window-shutter" style="--mdc-icon-size: 18px;"></ha-icon>
-            </ha-button>
-          </div>
-        ` : html`
-          <div class="cover-controls-wrapper">
-            <div class="cover-controls">
-              <span class="controls-label">${this._t("controls.cover_label")}</span>
-              <ha-button title=${this._t("controls.open")} @click=${() => this._onCoverCommand("open_cover")}>
-                <ha-icon icon="mdi:window-shutter-open" style="--mdc-icon-size: 18px;"></ha-icon>
-              </ha-button>
-              <ha-button title=${this._t("controls.stop")} @click=${() => this._onCoverCommand("stop_cover")}>
-                <ha-icon icon="mdi:stop" style="--mdc-icon-size: 18px;"></ha-icon>
-              </ha-button>
-              <ha-button title=${this._t("controls.close")} @click=${() => this._onCoverCommand("close_cover")}>
-                <ha-icon icon="mdi:window-shutter" style="--mdc-icon-size: 18px;"></ha-icon>
-              </ha-button>
-            </div>
-            <div class="cover-controls">
-              <span class="controls-label">${this._t("controls.tilt_label")}</span>
-              <ha-button title=${this._t("controls.tilt_open")} @click=${() => this._onCoverCommand("tilt_open")}>
-                <ha-icon icon="mdi:arrow-top-right" style="--mdc-icon-size: 18px;"></ha-icon>
-              </ha-button>
-              <ha-button title=${this._t("controls.tilt_stop")} @click=${() => this._onCoverCommand("tilt_stop")}>
-                <ha-icon icon="mdi:stop" style="--mdc-icon-size: 18px;"></ha-icon>
-              </ha-button>
-              <ha-button title=${this._t("controls.tilt_close")} @click=${() => this._onCoverCommand("tilt_close")}>
-                <ha-icon icon="mdi:arrow-bottom-left" style="--mdc-icon-size: 18px;"></ha-icon>
-              </ha-button>
-            </div>
-          </div>
-        `}
-        <div class="cal-form">
-          <div class="cal-field">
-            <select
-              class="ha-select"
-              id="position-select"
-              @change=${(e) => this._onPositionPresetChange(e.target.value)}
-            >
-              <option value="unknown" ?selected=${this._knownPosition === "unknown"}>${this._t("position.unknown")}</option>
-              <option value="open" ?selected=${this._knownPosition === "open"}>${this._t("position.open")}</option>
-              ${hasIndependentTilt
-                ? html`
-                    <option value="closed_tilt_open" ?selected=${this._knownPosition === "closed_tilt_open"}>${this._t("position.closed_tilt_open")}</option>
-                    <option value="closed_tilt_closed" ?selected=${this._knownPosition === "closed_tilt_closed"}>${this._t("position.closed_tilt_closed")}</option>
-                  `
-                : html`
-                    <option value="closed" ?selected=${this._knownPosition === "closed"}>${this._t("position.closed")}</option>
-                  `}
-            </select>
-          </div>
-        </div>
-      </div>
-    `;
-  }
-
-  _renderCalibration(calibrating) {
-    const state = this._getEntityState();
-    const attrs = state?.attributes || {};
-    const tiltMode = this._config?.tilt_mode || "none";
-    const hasTiltCalibration = tiltMode === "sequential_close" || tiltMode === "sequential_open" || tiltMode === "dual_motor" || tiltMode === "inline";
-
-    const availableAttributes = TIMING_ATTRIBUTES.filter(
-      ([key]) => {
-        if (!hasTiltCalibration && key.startsWith("tilt_")) return false;
-        return true;
-      }
-    );
-
-    const c = this._config;
-    const hasTravel = c?.travel_time_close || c?.travel_time_open;
-    const hasTilt = c?.tilt_time_close || c?.tilt_time_open;
-
-    const disabledKeys = new Set();
-    if (this._knownPosition === "unknown") {
-      availableAttributes.forEach(([key]) => disabledKeys.add(key));
-    } else if (this._knownPosition === "open") {
-      disabledKeys.add("travel_time_open");
-      disabledKeys.add("tilt_time_open");
-      if (hasTiltCalibration) {
-        // Tilt only changes when cover is closed — can't test from open
-        disabledKeys.add("tilt_time_close");
-        disabledKeys.add("tilt_startup_delay");
-      }
-    } else if (this._knownPosition === "closed") {
-      // Position closed (tilt matches)
-      disabledKeys.add("travel_time_close");
-      disabledKeys.add("tilt_time_close");
-    } else if (this._knownPosition === "closed_tilt_open") {
-      // tilt=100, cover closed. For sequential_close this is the "rest" state
-      // (slats at implicit-during-travel value), so travel_time_open and
-      // tilt_time_close are measurable. For sequential_open this is the
-      // "articulated" extreme (slats pushed open past cover-closed), so only
-      // tilt_time_close (restore slats to rest via motor up) is measurable.
-      disabledKeys.add("travel_time_close");
-      disabledKeys.add("tilt_time_open");
-      if (tiltMode === "sequential_open") {
-        disabledKeys.add("travel_time_open");
-        disabledKeys.add("travel_startup_delay");
-        disabledKeys.add("min_movement_time");
-      }
-    } else if (this._knownPosition === "closed_tilt_closed") {
-      // tilt=0, cover closed. For sequential_close this is the "articulated"
-      // extreme — only tilt_time_open (restore slats) is measurable. For
-      // sequential_open this is the "rest" state (slats at implicit), so
-      // travel_time_open and tilt_time_open (articulate further down) are
-      // measurable.
-      disabledKeys.add("travel_time_close");
-      disabledKeys.add("tilt_time_close");
-      if (tiltMode !== "sequential_open") {
-        disabledKeys.add("travel_time_open");
-        disabledKeys.add("travel_startup_delay");
-        disabledKeys.add("min_movement_time");
-      }
-    }
-
-    // Startup delay requires the corresponding time to be calibrated first
-    if (!hasTravel) disabledKeys.add("travel_startup_delay");
-    if (!hasTilt) disabledKeys.add("tilt_startup_delay");
-    if (!hasTravel) disabledKeys.add("min_movement_time");
-
-    if (calibrating) {
-      const calAttr = attrs.calibration_attribute || this._calibratingAttribute;
-      const calLabel = this._t(`timing.${calAttr}`);
-      return html`
-        <div class="section calibration-active">
-          <div class="field-label cal-label">
-            <ha-icon icon="mdi:tune" style="--mdc-icon-size: 20px;"></ha-icon>
-            ${this._t("calibration.active")}
-          </div>
-          <div class="cal-active-body">
-            <strong>${calLabel}</strong>
-            ${attrs.calibration_final_step
-              ? html`<span class="cal-step"
-                  >${this._t("calibration.final_step")}</span
-                >`
-              : attrs.calibration_step
-                ? html`<span class="cal-step"
-                    >${this._t("calibration.step", { step: attrs.calibration_step })}</span
-                  >`
-                : ""}
-            <div class="cal-active-buttons">
-              <ha-button @click=${() => this._onStopCalibration(true)}
-                >${this._t("calibration.cancel")}</ha-button
-              >
-              <ha-button unelevated @click=${() => this._onStopCalibration(false)}
-                >${this._t("calibration.finish")}</ha-button
-              >
-            </div>
-          </div>
-        </div>
-      `;
-    }
-
-    return html`
-      <div class="section">
-        <div class="field-label">${this._t("calibration.label")}</div>
-        <div class="cal-form">
-          <div class="cal-field">
-            <label class="sub-label" for="cal-attribute">${this._t("calibration.attribute_label")}</label>
-            <select class="ha-select" id="cal-attribute"
-              @change=${() => this.requestUpdate()}
-            >
-              ${availableAttributes.map(
-                ([key, labelKey]) =>
-                  html`<option value=${key} ?disabled=${disabledKeys.has(key)}>${this._t(labelKey)}</option>`
-              )}
-            </select>
-          </div>
-          <ha-button unelevated ?disabled=${this._knownPosition === "unknown"} @click=${this._onStartCalibration}
-            >${this._t("calibration.start")}</ha-button
-          >
-        </div>
-        ${this._knownPosition === "unknown"
-          ? html`<div class="helper-text" style="margin-top: 8px;">
-              ${this._t("calibration.set_position_first")}
-            </div>`
-          : html`<div class="helper-text" style="margin-top: 8px;">
-              ${this._getCalibrationHint()}
-            </div>`}
-      </div>
-    `;
   }
 
   // --- Styles ---
